+++ NEWS.md
@@ -1,5 +1,9 @@
 # FSA 0.8.32 Ongoing
-* `hist.formula()`: Fixed bug with y-axes when `freq=FALSE` is used (fixes [#62](https://github.com/droglenc/FSA/issues/62); thanks to @carlostorrescubila).
+* Moved a bunch of  plotting examples in the documentation to `tests\plottests\` to speed up testing. The `tests\plottests\` was added to `.Rbuildignore` .
+* `hist.formula()`: Modified. Fixed bug with y-axes when `freq=FALSE` is used (fixes [#62](https://github.com/droglenc/FSA/issues/62); thanks to @carlostorrescubila).
+* `fitPlot()`: Modified. Fixed bugs with handling models that used character rather than factor variables.
+* `psdPlot()`: Modified. Fixed bug with box around the plot when `add.psd=FALSE`. Added 5% expansion to top of y-axis so that bars did not run into the box.
+* `residPlot()`: Modified. Fixed bugs with handling models that used character rather than factor variables.
 
 # FSA 0.8.31 7-Nov20
 * Now using roxygen v7.1.1.